Extract test debot interfaces to a separate file
